requirements.txt
redis
python-dotenv
diffusers
matplotlib